refactor(providers): move hourCycle from l10n to a global F0Provider user setting
Addresses review feedback: 12h/24h is a per-user platform preference (set
once, e.g. Factorial's Preferences > Time format), not a locale-derived
value, so it shouldn't live in the l10n context nor be passed per component.

- Add `hourCycle` as a global user setting on F0Provider, threaded through
  UserPlatformProvider and exposed via `useHourCycle()`.
- Move the `HourCycle` type to providers/user-platafform.
- Revert the L10nContextValue / l10n defaults changes.
- TimeFieldRenderer now reads `useHourCycle()`; unset still falls back to the
  native input (opt-in preserved).
- Update the story to configure hourCycle via the provider.
