+++ packages/react/src/lib/providers/l10n/index.ts
@@ -1,4 +1,3 @@
 export * from "./l10n-provider"
 export * from "./l10n-provider-defaults"
 export * from "./use-date-fns-locale"
-export * from "./types"
