+++ packages/react/src/lib/providers/l10n/l10n-provider-defaults.ts
@@ -7,7 +7,4 @@ export const defaults: L10nContextValue = {
   date: {
     weekStartsOn: WeekStartDay.Monday,
   },
-  time: {
-    hourCycle: "24h",
-  },
 }
