+++ packages/react/src/lib/providers/user-platafform/types.ts
@@ -1,1 +1,4 @@
 export type Platform = "mac" | "windows" | "linux" | "mobile" | "unknown"
+
+/** How times are displayed and edited: 12-hour (AM/PM) or 24-hour. */
+export type HourCycle = "12h" | "24h"
